extract main collection items
